Finished up initial scraper of match pages
